+++ api/CHANGELOG.md
@@ -13,6 +13,7 @@
 - Aligned the pencil button on the patch bundles tab to the other buttons (along with giving it the same size as the other buttons)
 - Added a search button next to all version listings under "Show suggested versions" on the app selection page. Tapping the new button searches with google the package name and the version number (example: com.google.android.youtube 20.51.38)
 - Redesigned the UI of the app selection page
+- Patch filter selections on the patch selection page now persist https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/73
 
 
 # Bug fixes